refactor: add view decorators to inject player
